conversion of interactions with zone into a queued, periodic set of tasks
