+++ src/main/scala/net/psforever/objects/vehicles/VehicleControl.scala
@@ -707,7 +707,7 @@ class VehicleControl(vehicle: Vehicle)
   /**
     * Tell the given targets that
     * water causes vehicles to become disabled if they dive off too far, too deep.
-    * @see `InteractWithEnvironment`
+    * @see `InteractingWithEnvironment`
     * @see `OxygenState`
     * @see `OxygenStateTarget`
     * @param percentage the progress bar completion state
@@ -721,7 +721,7 @@ class VehicleControl(vehicle: Vehicle)
                                      ): Unit = {
     val vtarget = Some(OxygenStateTarget(vehicle.GUID, OxygenState.Suffocation, percentage))
     targets.foreach { target =>
-      target.Actor ! InteractWithEnvironment(target, body, vtarget)
+      target.Actor ! InteractingWithEnvironment(target, body, vtarget)
     }
   }
 
@@ -745,7 +745,7 @@ class VehicleControl(vehicle: Vehicle)
       //keep doing damage
       if (vehicle.Health > 0) {
         import scala.concurrent.ExecutionContext.Implicits.global
-        interactionTimer = context.system.scheduler.scheduleOnce(delay = 250 milliseconds, self, InteractWithEnvironment(obj, body, None))
+        interactionTimer = context.system.scheduler.scheduleOnce(delay = 250 milliseconds, self, InteractingWithEnvironment(obj, body, None))
       }
     }
   }
